Added DataController.

+++ sources/radio/help_structures.h
@@ -6,7 +6,7 @@
 #include <utility>
 #include <vector>
 
-using Frequency = int64_t;
+using Frequency = int32_t;
 using FrequencyRange = std::pair<Frequency, Frequency>;
 using FrequencyFlush = std::pair<Frequency, bool>;
 using TransmissionNotification = Notification<std::vector<FrequencyFlush>>;
